requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
faiss-cpu>=1.7.4
pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pandas>=2.0.0
timm>=0.9.2
pyyaml>=6.0
